+++ mobile/android/android-components/components/feature/customtabs/src/test/java/mozilla/components/feature/customtabs/CustomTabIntentProcessorTest.kt
@@ -5,6 +5,8 @@
 package mozilla.components.feature.customtabs
 
 import android.content.Intent
+import android.os.Bundle
+import android.provider.Browser
 import androidx.browser.customtabs.CustomTabsIntent
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -21,6 +23,7 @@ import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import mozilla.components.support.test.whenever
+import mozilla.components.support.utils.toSafeIntent
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertFalse
@@ -65,6 +68,41 @@ class CustomTabIntentProcessorTest {
         handler.process(intent)
         verify(sessionManager).add(anySession(), eq(false), eq(null), eq(null), eq(null))
         verify(engineSession).loadUrl("http://mozilla.org", flags = LoadUrlFlags.external())
+        verify(intent).putExtra(eq(EXTRA_SESSION_ID), any<String>())
+
+        val customTabSession = sessionManager.all[0]
+        assertNotNull(customTabSession)
+        assertEquals("http://mozilla.org", customTabSession.url)
+        assertEquals(Source.CUSTOM_TAB, customTabSession.source)
+        assertNotNull(customTabSession.customTabConfig)
+        assertFalse(customTabSession.private)
+    }
+
+    @Test
+    fun processCustomTabIntentWithAdditionalHeaders() = runBlockingTest {
+        val engine = mock<Engine>()
+        val sessionManager = spy(SessionManager(engine))
+        doReturn(engineSession).`when`(sessionManager).getOrCreateEngineSession(anySession())
+        val useCases = SessionUseCases(sessionManager)
+
+        val handler =
+                CustomTabIntentProcessor(sessionManager, useCases.loadUrl, testContext.resources)
+
+        val intent = mock<Intent>()
+        whenever(intent.action).thenReturn(Intent.ACTION_VIEW)
+        whenever(intent.hasExtra(CustomTabsIntent.EXTRA_SESSION)).thenReturn(true)
+        whenever(intent.dataString).thenReturn("http://mozilla.org")
+        whenever(intent.putExtra(any<String>(), any<String>())).thenReturn(intent)
+
+        val headersBundle = Bundle().apply {
+            putString("X-Extra-Header", "true")
+        }
+        whenever(intent.getBundleExtra(Browser.EXTRA_HEADERS)).thenReturn(headersBundle)
+        val headers = handler.getAdditionalHeaders(intent.toSafeIntent())
+
+        handler.process(intent)
+        verify(sessionManager).add(anySession(), eq(false), eq(null), eq(null), eq(null))
+        verify(engineSession).loadUrl("http://mozilla.org", flags = LoadUrlFlags.external(), additionalHeaders = headers)
         verify(intent).putExtra(eq(EXTRA_SESSION_ID), any<String>())
 
         val customTabSession = sessionManager.all[0]